improve input handling

+++ Automata/Core/ComponentSystem.cs
@@ -9,11 +9,6 @@ namespace Automata.Core
     public abstract class ComponentSystem
     {
         /// <summary>
-        ///     The <see cref="SystemManager" /> uses this boolean to track whether it has enabled the given System.
-        /// </summary>
-        // public bool IsEnabled { get; set; } = false;
-
-        /// <summary>
         ///     This is a hint to the <see cref="SystemManager" /> as to what types the system uses for updates.
         ///     If the types aren't present in the <see cref="EntityManager" />, the system's <see cref="Update" /> will be
         ///     skipped.
@@ -26,21 +21,13 @@ namespace Automata.Core
         public virtual void Registered() { }
 
         /// <summary>
-        ///     Method called when the <see cref="ComponentSystem" /> is enabled.
-        /// </summary>
-        /// <remarks>
-        ///     Enabling happens before the first execution of <see cref="Update" />.
-        /// </remarks>
-        // public virtual void Enabled(EntityManager entityManager) { }
-
-        /// <summary>
         ///     Method called once per frame.
         /// </summary>
         /// <remarks>
         ///     If none of the types in <see cref="UtilizedComponentTypes" /> are active in the <see cref="EntityManager" />, this
         ///     method is skipped.
         /// </remarks>
-        public virtual void Update(EntityManager entityManager, double deltaTime) { }
+        public virtual void Update(EntityManager entityManager, float deltaTime) { }
 
         /// <summary>
         ///     Method called when the system is destroyed.
